Enhances `DI_DO_action` and `inverted_DI_DO_action` for OR logic (#1035)
* Initial plan

* Implement multiple_DI_DO_action for OR logic with multiple DI signals

* Fix test failures by adding missing step creation and enum entry

* Fix action array indexing by moving multiple_DI_DO_action to correct position

* Add comprehensive test coverage for multiple_DI_DO_action to improve code coverage

* Convert raw for-loops to range-based for-loops for better C++ style

* Add device ordering validation for DI signals before DO signals

* Integrate multiple DI functionality into existing DI_DO_action and inverted_DI_DO_action classes

* Removes redundant DI/DO action tests

* Adds device order check test

* Prevents accidental modification of device list

Ensures that the device list in `DI_DO_action` is not accidentally modified
by declaring it as a const reference in both `evaluate()` and `finalize()` methods.

* Refactors DI/DO device type checking

Simplifies DI/DO action device type checks by introducing a dedicated function for determining DI device types.
This improves code readability and maintainability by centralizing the logic for identifying DI devices.

---------

+++ common/PAC-driver/g_errors.cpp
@@ -62,9 +62,11 @@ int tech_dev_error::save_as_Lua_str( char *str )
         res += fmt::format_to_n( str + res, MAX_COPY_SIZE,
             "id_n={},\n", simple_device->get_serial_n() ).size;
         res += fmt::format_to_n( str + res, MAX_COPY_SIZE,
-            "id_object_alarm_number={},\n", -simple_device->get_error_id() ).size;
+            "id_object_alarm_number={},\n",
+            -simple_device->get_error_id() ).size;
         res += fmt::format_to_n( str + res, MAX_COPY_SIZE,
-            "id_type={},\n", simple_device->get_type() ).size;
+            "id_type={},\n",
+            static_cast<int>( simple_device->get_type() ) ).size;
 
         res += fmt::format_to_n( str + res, MAX_COPY_SIZE,
             "suppress={}\n", alarm_params & P_IS_SUPPRESS ? "true" : "false" ).size;
